+++ topologic/__init__.py
@@ -11,18 +11,12 @@ from .assertions import assert_is_graph, \
     assert_single_connected_components
 
 # load individual modules into the top level topologic namespace
-from .louvain import best_partition, \
-    induce_graph_by_communities, \
-    induced_community_graph, \
-    induced_graph_for_best_partition, \
-    InducedCommunityGraph
 from .connected_components import number_connected_components, \
     largest_connected_component, \
     connected_components_generator
 from .partitioned_graph import PartitionedGraph
 from .graph_augmentation import rank_edges, self_loop_augmentation
 from .eigenvalue_elbows import get_elbows_from_eigenvalues
-from .modularity import q_score
 from .distance import cosine_distance, euclidean_distance, mahalanobis_distance
 
 from .io.bipartite_graph_consolidator import consolidate_bipartite
@@ -37,11 +31,11 @@ from . import projection
 from . import iterators
 from . import statistics
 from . import embedding
+from . import partitioning
 
 __all__ = [
     'DialectException',
     'GraphProperties',
-    'InducedCommunityGraph',
     'InvalidGraphError',
     'PartitionedGraph',
     'PotentialEdgeColumnPair',
@@ -49,7 +43,6 @@ __all__ = [
     'assert_is_weighted',
     'assert_is_weighted_graph',
     'assert_single_connected_components',
-    'best_partition',
     'calculate_ari_scores',
     'calculate_ari',
     'connected_components_generator',
@@ -59,13 +52,9 @@ __all__ = [
     'find_edges',
     'find_elbows',
     'get_elbows_from_eigenvalues',
-    'induce_graph_by_communities',
-    'induced_community_graph',
-    'induced_graph_for_best_partition',
     'largest_connected_component',
     'mahalanobis_distance',
     'number_connected_components',
-    'q_score',
     'self_loop_augmentation',
     'UnweightedGraphError'
 ]
