+++ topologic/partitioning/louvain_stub.py
@@ -1,0 +1,8 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+import community
+
+
+# Documentation note: see docs/topologic.partitioning.rst for document string to louvain function alias
+louvain = community.best_partition
